+++ NEWS.md
@@ -15,6 +15,10 @@
   
   These dynamic values may be useful for validations that get expanded into multiple steps.
 
+## Minor improvements and bug fixes
+
+* When no columns are returned from a `{tidyselect}` expression in `columns`, the agent's report now displays the originally supplied *expression* instead of simply blank (e.g., in `create_agent(small_table) %>% col_vals_null(matches("z"))`).
+
 # pointblank 0.11.4
 
 * Fixes issue with gt `0.9.0` compatibility.